[ObservableObject] fire an observer when properties or children change (#1)
This change adds the `ObservableObject` - the root of this entire package. It fires and observer callback when a property (or sub-property) changes values. It does not fire of the value is the same as previously.

```js
/**
 * observer callback for when a property changes on an ObservableObject
 * @callback observableObjectObserver
 * @param {string} value - the value of the property that changed on the ObservableObject
 * @param {string} propertyChanged - the name of the property that changed on the ObservableObject
 * @param {object} object - the containing object which had the change
 */
const observer = (value, propertyChanged, object) => 
  console.log(`ChangeObserved::(${value}, ${propertyChanged}, ${object})`)

const obj = new ObservableObject(observer)
obj.test = true  // console.log => "ChangeObserved::(true, test, [object Object])"
```